requirements.txt
flask==2.2.3
torch==2.0.1
transformers==4.30.2
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.24.3
werkzeug==2.2.3
accelerate==0.20.3
sentencepiece==0.1.99